Rip out copying code to data segment
The biggest change is that in addition to polymorphism, it was also done
because the interrupt handler only has CS set, not DS. We can simply
restore DS however, and pop it back to what it was right before IRET.

This also changes how some variables were addressed in the data segment,
as well as the address of the soft interrupt handler. The variables are
slightly annoying because they need to be adapted to live in ps2's part
of the data segment.

This also removes a lot of indirection (for example: the interrupt
handling and the enable/disable polymorphism), and basically solidifies
us as supporting VMware only. No harm in writing another driver for
another platform...