Dirty hack for working wheel support
It turns out we can actually call Windows functions! We're a 16-bit
driver living in the same address space. What is trickier is IIRC we can
only link against KERNEL. However, some drivers (i.e. COMM or network)
are chatty and can use MessageBox or PostMessage to interact with the
user. They're just libraries, after all, so they can talk to the rest of
the world.

We use this trick to send a WM_VSCROLL (line up/down) to the currently
focused window. However, not all windows are focusable (i.e. winhelp),
so it would help to be able to get what's under the cursor; a solution
ahs been identified.

Comments will get cleaned up later...