Look for window under mouse, calling convention
It seems these calls are *mostly* safe to call from an interrupt
handler. (Chen documents PostMessage as definitely safe.)

The new tactic is instead of using the focused window (which the thing
you want to scroll may not necessarily be focusable), use the window
under the cursor, which has the same semantics on modern OSes. The
advantage is we can get the HWND of non-active windows too, which
cements the fact it works like a modern OS too (It took until Windows 10
for wheel messages to reach inactive windows!).

Some programs handle this great (i.e. WinHelp), others seemingly
confusingly (wheel works on the non-client area of Write, but not the
document itself), or wonky (Program Manager's MDI area gets confused).
It may need some per-app tweaks or maybe I'm missing something for
better handling. For per-app tweaks, we could consider looking at the
window class.

However, there still are some occasional crashes. Not sure what causes
these other than maybe excessive usage? Help debugging would be
appreciatyed.

To make things easier on myself, I've changed the calling convention of
the function to use the C macros MS provides and occasionally uses. This
makes parameter and local variable on a stack frame easier to handle,
though I'm tempted to revert back some of it because it is a bit ugly,
especially in SoftICE where MASM had a hard time making symbols for
cProcs.

The wheel handling is also moved much later, after the event procedure.
I think this is me trying to play it safe.

@stsp, this is perhaps interesting for you, but I would be hesitant with
my current implementation.