Make it so the wheel is behind an INI option
I want to merge this branch, have the feature available, but warn people
it might be buggy (plus give others like dosemu2 et al a chance to
figure out a way to make this some kind of standard). So you get a
feature flag. A UI to expose it would be nice, but for now...

If you want to enable it (be warned: maybe bugs), in SYSTEM.INI:

[VMwareMouse]
EnableWheel=1